feat(catalog): make content header customizable
