+++ docs/features/software-catalog/catalog-customization.md
@@ -416,6 +416,29 @@ const routes = (
 
 The same method can be used to customize the _default_ filters with a different interface - for such usage, the generic argument isn't needed since the filter shape remains the same as the default.
 
+## Customize Content Header
+
+The `CatalogIndexPage` comes with a default content header, but you may want to customize it:
+
+```tsx title="packages/app/src/App.tsx"
+import { DismissableBanner } from '@backstage/core-components';
+
+<Route
+  path="/catalog"
+  element={
+    <CatalogIndexPage
+      contentHeader={
+        <DismissableBanner
+          id="repository_migration_notice"
+          variant="info"
+          message="We're currently migrating repositories to a new provider. Your entities might be temporarily missing from the catalog."
+        />
+      }
+    />
+  }
+/>;
+```
+
 ## Advanced Customization
 
 For those where none of the above fits their needs you can take the option of creating a fully custom `CatalogIndexPage`.
